+++ packages/web-shared/src/components/new-trace-viewer/components/span-markers.tsx
@@ -2,7 +2,7 @@
 
 import { ArrowLeft, ArrowRight } from 'lucide-react';
 import type { ReactNode } from 'react';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../lib/cn';
 import { TimestampTooltip } from '../../ui/timestamp-tooltip';
 import type { SpanMarker } from '../utils';
