+++ packages/web-shared/src/components/ui/collapsible.tsx
@@ -8,7 +8,7 @@ import {
   useContext,
   useState,
 } from 'react';
-import { cn } from '../../lib/utils';
+import { cn } from '../../lib/cn';
 
 type CollapsibleVariant = 'section' | 'card';
